Test and implement Evaluatable interface on Symbol.
The Symbol class implements the Evaluatable class and tests have been
writen to show that Evaluatable method calls on Symbol instances produce
the results that will be expected by an Evaluator.

+++ src/main/java/Evaluator/Evaluator.java
@@ -2,7 +2,7 @@ public abstract class Evaluator {
 	ExpressionTree expressionTree;
 	
 	// get the expression tree from the context.
-	public float evaluate(ExpressionTree expressionTree){
+	public double evaluate(ExpressionTree expressionTree){
 		if (expressionTree == null)
 			throw new NullPointerException(
 				"An ExpressionTree is required.");
@@ -10,5 +10,5 @@ public abstract class Evaluator {
 		return evaluate();
 	}
 	
-	protected abstract float evaluate();
+	protected abstract double evaluate();
 }
